test: consolidate test suite (Phase 6+ revised)
Consolidate test files to reduce fragmentation while preserving all tests:

CLI tests (13 → 9 files, -4):
- Merged test_cli_completions.py, test_cli_mcp.py, test_cli_reset.py,
  test_cli_serve.py into test_cli_commands.py
- Preserved both subprocess integration tests AND CliRunner unit tests
- Clear section organization with === banners

Importer tests (7 → 5 files, -2):
- Created test_importers_unit.py consolidating ChatGPT, Claude, Codex
  provider-specific unit tests
- Kept test_importers_parametrized.py (cross-provider integration)
- Kept test_importers_properties.py (Hypothesis property-based)

Results:
- Total test files: 96 → 90 (-6)
- Test count unchanged: 2125 tests passing
- No coverage loss
