--- tests/test_cli_mcp.py
@@ -1,108 +1,0 @@
-"""Tests for polylogue mcp command.
-
-Coverage targets:
-- mcp_command: MCP server startup
-- --transport: Transport type selection
-- Error handling for missing dependencies
-"""
-
-from __future__ import annotations
-
-from unittest.mock import MagicMock, patch
-
-import pytest
-from click.testing import CliRunner
-
-from polylogue.cli.commands.mcp import mcp_command
-
-
-@pytest.fixture
-def runner():
-    """CLI runner for testing."""
-    return CliRunner()
-
-
-@pytest.fixture
-def mock_env():
-    """Create mock AppEnv for tests."""
-    mock_ui = MagicMock()
-    mock_ui.plain = True
-    mock_ui.console = MagicMock()
-
-    env = MagicMock()
-    env.ui = mock_ui
-    return env
-
-
-class TestMcpCommand:
-    """Tests for the mcp command."""
-
-    def test_default_transport_is_stdio(self, runner, mock_env):
-        """Default transport is stdio."""
-        with patch("polylogue.mcp.server.serve_stdio") as mock_serve:
-            result = runner.invoke(mcp_command, [], obj=mock_env)
-
-            # Should call serve_stdio
-            mock_serve.assert_called_once()
-            assert result.exit_code == 0
-
-    def test_explicit_stdio_transport_works(self, runner, mock_env):
-        """--transport stdio works."""
-        with patch("polylogue.mcp.server.serve_stdio") as mock_serve:
-            result = runner.invoke(mcp_command, ["--transport", "stdio"], obj=mock_env)
-
-            mock_serve.assert_called_once()
-            assert result.exit_code == 0
-
-    def test_missing_mcp_dependencies_error(self, runner, mock_env):
-        """Missing MCP dependencies show helpful error."""
-        # Patch the import to raise ImportError
-        import sys
-
-        with patch.dict(sys.modules, {"polylogue.mcp.server": None}):
-            # Force ImportError by patching the actual import
-            def mock_import(*args, **kwargs):
-                raise ImportError("No module named 'mcp'")
-
-            with patch("builtins.__import__", side_effect=mock_import):
-                result = runner.invoke(mcp_command, [], obj=mock_env)
-
-                # Should fail with helpful message
-                assert result.exit_code != 0 or mock_env.ui.console.print.called
-
-    def test_unsupported_transport_error(self, runner, mock_env):
-        """Unsupported transport type raises error."""
-        # The Click choice validation should reject this
-        from polylogue.cli.click_app import cli
-
-        result = runner.invoke(cli, ["mcp", "--transport", "http"])
-
-        assert result.exit_code != 0
-
-    def test_mcp_help_shows_description(self, runner):
-        """MCP help shows useful description."""
-        from polylogue.cli.click_app import cli
-
-        result = runner.invoke(cli, ["mcp", "--help"])
-
-        assert result.exit_code == 0
-        assert "mcp" in result.output.lower()
-        assert "server" in result.output.lower() or "protocol" in result.output.lower()
-
-
-class TestMcpServerIntegration:
-    """Integration tests for MCP server (when dependencies are available)."""
-
-    def test_serve_stdio_can_be_imported(self):
-        """serve_stdio can be imported if mcp is installed."""
-        try:
-            from polylogue.mcp.server import serve_stdio
-            assert callable(serve_stdio)
-        except ImportError:
-            # MCP not installed, skip
-            pytest.skip("MCP dependencies not installed")
-
-    def test_mcp_server_module_exists(self):
-        """MCP server module exists in package."""
-        import polylogue.mcp as mcp_module
-        assert hasattr(mcp_module, "__file__")
